Refactor around trade legality (#35)
* Filter trades as they are executed

Previously, we checked that all proposed trades were legal before we executed any of them.

This is a problem when we're planning trades ahead of time. Some trades are not legal yet, but will be by the time we are ready to place them.

In this commit, we do not check that a trade is legal until we have already placed the trades that come before it.

Mutating object all happens in the abstract class

Update LiveMarketAdapter

back to old mutation strategy

Fixed edge case where balances were being removed when estimating value

* No longer mutating stuff in ProposedTrade

Previously, if we tried to set_bid_amount() greater than we currently had, ProposedTrade would limit us. Really, we want the MarketAdapter to do that. We may not be able to afford a trade at the moment we propose it, but by the time that trade gets executed, we might be able to make it.

* Fixed tests

* Added check for holding 0 balance

* Less side-effecty business

* MarketAdapter and Strategy no longer share MarketState

Previously, we passed the MarketAdapter's market_state to Strategy::propose_trades(). If the strategy were to mutate market_state (even accidentally), it could pull some shennanigans, especially in the market adapters (think artificially inflating balances of particular coins).

In addition, MarketAdapters kept their balance in two places: one, in the market_state, and another, in their balances property.

This commit fixes both of these issues. First, we roll MarketAdapter's balances property into the market_state, and read it only from this property. Second, we make a copy of the MarketAdapter.market_state property before we feed it to Strategy.propose_trades()